# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Programme.lv backend
    description: Service for managing users, tasks, and submissions.
    version: 0.0.1
host: localhost:80
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /submissions:
        get:
            tags:
                - submissions
            summary: listSubmissions submissions
            description: List all submissions
            operationId: submissions#listSubmissions
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/SubmissionResponse'
                "400":
                    description: Bad Request response.
                    schema:
                        type: string
                "401":
                    description: Unauthorized response.
                    schema:
                        type: string
                "404":
                    description: Not Found response.
                    schema:
                        type: string
                "500":
                    description: Internal Server Error response.
                    schema:
                        type: string
            schemes:
                - http
        post:
            tags:
                - submissions
            summary: createSubmission submissions
            description: Create a new submission
            operationId: submissions#createSubmission
            parameters:
                - name: CreateSubmissionRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/SubmissionsCreateSubmissionRequestBody'
                    required:
                        - uuid
                        - submission
                        - username
                        - createdAt
                        - evaluation
                        - language
                        - task
            responses:
                "201":
                    description: Created response.
                    schema:
                        $ref: '#/definitions/SubmissionsCreateSubmissionResponseBody'
                        required:
                            - uuid
                            - submission
                            - username
                            - createdAt
                            - evaluation
                            - language
                            - task
                "400":
                    description: Bad Request response.
                    schema:
                        type: string
                "401":
                    description: Unauthorized response.
                    schema:
                        type: string
                "404":
                    description: Not Found response.
                    schema:
                        type: string
                "500":
                    description: Internal Server Error response.
                    schema:
                        type: string
            schemes:
                - http
    /submissions/{uuid}:
        get:
            tags:
                - submissions
            summary: getSubmission submissions
            description: Get a submission by UUID
            operationId: submissions#getSubmission
            parameters:
                - name: uuid
                  in: path
                  description: UUID of the submission
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/SubmissionsGetSubmissionResponseBody'
                        required:
                            - uuid
                            - submission
                            - username
                            - createdAt
                            - evaluation
                            - language
                            - task
                "400":
                    description: Bad Request response.
                    schema:
                        type: string
                "401":
                    description: Unauthorized response.
                    schema:
                        type: string
                "404":
                    description: Not Found response.
                    schema:
                        type: string
                "500":
                    description: Internal Server Error response.
                    schema:
                        type: string
            schemes:
                - http
...
definitions:
    EvaluationRequestBody:
        title: EvaluationRequestBody
        type: object
        properties:
            possibleScore:
                type: integer
                description: Possible score of the evaluation
                example: 100
                format: int64
            receivedScore:
                type: integer
                description: Received score of the evaluation
                example: 85
                format: int64
            status:
                type: string
                description: Status of the evaluation
                example: completed
            uuid:
                type: string
                description: UUID of the evaluation
                example: 123e4567-e89b-12d3-a456-426614174000
        description: Represents the evaluation of a submission
        example:
            possibleScore: 100
            receivedScore: 85
            status: completed
            uuid: 123e4567-e89b-12d3-a456-426614174000
        required:
            - uuid
            - status
            - receivedScore
            - possibleScore
    EvaluationResponse:
        title: EvaluationResponse
        type: object
        properties:
            possibleScore:
                type: integer
                description: Possible score of the evaluation
                example: 100
                format: int64
            receivedScore:
                type: integer
                description: Received score of the evaluation
                example: 85
                format: int64
            status:
                type: string
                description: Status of the evaluation
                example: completed
            uuid:
                type: string
                description: UUID of the evaluation
                example: 123e4567-e89b-12d3-a456-426614174000
        description: Represents the evaluation of a submission
        example:
            possibleScore: 100
            receivedScore: 85
            status: completed
            uuid: 123e4567-e89b-12d3-a456-426614174000
        required:
            - uuid
            - status
            - receivedScore
            - possibleScore
...
    ProgrammingLangRequestBody:
        title: ProgrammingLangRequestBody
        type: object
        properties:
            fullName:
                type: string
                description: Full name of the programming language
                example: Go
            id:
                type: string
                description: ID of the programming language
                example: go
            monacoId:
                type: string
                description: Monaco editor ID for the programming language
                example: go
        description: Represents a programming language
        example:
            fullName: Go
            id: go
            monacoId: go
        required:
            - id
            - fullName
            - monacoId
    ProgrammingLangResponse:
        title: ProgrammingLangResponse
        type: object
        properties:
            fullName:
                type: string
                description: Full name of the programming language
                example: Go
            id:
                type: string
                description: ID of the programming language
                example: go
            monacoId:
                type: string
                description: Monaco editor ID for the programming language
                example: go
        description: Represents a programming language
        example:
            fullName: Go
            id: go
            monacoId: go
        required:
            - id
            - fullName
            - monacoId
...
    SubmissionResponse:
        title: SubmissionResponse
        type: object
        properties:
            createdAt:
                type: string
                description: Creation date of the submission
                example: "2024-08-08T10:30:00Z"
                format: date-time
            evaluation:
                $ref: '#/definitions/EvaluationResponse'
            language:
                $ref: '#/definitions/ProgrammingLangResponse'
            submission:
                type: string
                description: The code submission
                example: print(factorial(5))
            task:
                $ref: '#/definitions/TaskResponse'
            username:
                type: string
                description: Username of the user who submitted
                example: coder123
            uuid:
                type: string
                description: UUID of the submission
                example: 123e4567-e89b-12d3-a456-426614174000
        description: Represents a submission
        example:
            createdAt: "2024-08-08T10:30:00Z"
            evaluation:
                possibleScore: 100
                receivedScore: 85
                status: completed
                uuid: 123e4567-e89b-12d3-a456-426614174000
            language:
                fullName: Go
                id: go
                monacoId: go
            submission: print(factorial(5))
            task:
                cpu_time_limit_seconds: 0.5
                default_md_statement:
                    input: Ievaddatu pirmajā rindā dotas trīs naturālu skaitļu...
                    notes: Recusandae dicta qui.
                    output: Izvaddatu vienīgajā rindā jābūt veselam skaitlim...
                    scoring: Aut non maiores voluptatem alias maxime ipsum.
                    story: Krišjānis ir uzkonstruējis kvadrātveida putekļsūcēju (saīsināti – KP), kas ir neaizstājams palīgs viņa darbnīcas uzkopšanā...
                default_pdf_statement_url: https://dvhk4hiwp1rmf.cloudfront.net/task-pdf-statements/f27386f1b69f3c020335fe7c84316c5da099933832043e6a5820bcdb0cd66a81.pdf
                difficulty_rating: 3
                examples:
                    - input: |
                        5 9 3
                        A....X..B
                        ..X..X.X.
                        .XXX.XX..
                        X.X.X..X.
                        ...XX....
                      md_note: Alias enim corporis est.
                      output: |
                        10
                    - input: |
                        5 9 3
                        A....X..B
                        ..X..X.X.
                        .XXX.XX..
                        X.X.X..X.
                        ...XX....
                      md_note: Alias enim corporis est.
                      output: |
                        10
                    - input: |
                        5 9 3
                        A....X..B
                        ..X..X.X.
                        .XXX.XX..
                        X.X.X..X.
                        ...XX....
                      md_note: Alias enim corporis est.
                      output: |
                        10
                illustration_img_url: https://dvhk4hiwp1rmf.cloudfront.net/task-illustrations/bafcc0aa1b4c56faa44f5f3b2a5abd9529af941eb9a9b10f541b436762d4948a.png
                memory_limit_megabytes: 256
                origin_notes:
                    lv: Uzdevums parādījās Latvijas 37. informātikas olimpiādes (2023./2024. gads) skolas kārtā.
                origin_olympiad: LIO
                published_task_id: kvadrputekl
                task_full_name: Kvadrātveida putekļsūcējs
                visible_input_subtasks:
                    - inputs:
                        - |
                          5 5 3
                          A.X..
                          X.B.X
                          ....X
                          XX...
                          X.XXX
                        - |
                          6 3 1
                          ...
                          XA.
                          .X.
                          X..
                          B..
                          ...
                        - |
                          6 4 2
                          X...
                          .AXX
                          X..X
                          .X.X
                          XXXX
                          X.BX
                      subtask: 1
                    - inputs:
                        - |
                          5 5 3
                          A.X..
                          X.B.X
                          ....X
                          XX...
                          X.XXX
                        - |
                          6 3 1
                          ...
                          XA.
                          .X.
                          X..
                          B..
                          ...
                        - |
                          6 4 2
                          X...
                          .AXX
                          X..X
                          .X.X
                          XXXX
                          X.BX
                      subtask: 1
                    - inputs:
                        - |
                          5 5 3
                          A.X..
                          X.B.X
                          ....X
                          XX...
                          X.XXX
                        - |
                          6 3 1
                          ...
                          XA.
                          .X.
                          X..
                          B..
                          ...
                        - |
                          6 4 2
                          X...
                          .AXX
                          X..X
                          .X.X
                          XXXX
                          X.BX
                      subtask: 1
            username: coder123
            uuid: 123e4567-e89b-12d3-a456-426614174000
        required:
            - uuid
            - submission
            - username
            - createdAt
            - evaluation
            - language
            - task
    SubmissionsCreateSubmissionRequestBody:
        title: SubmissionsCreateSubmissionRequestBody
        type: object
        properties:
            createdAt:
                type: string
                description: Creation date of the submission
                example: "2024-08-08T10:30:00Z"
                format: date-time
            evaluation:
                $ref: '#/definitions/EvaluationRequestBody'
            language:
                $ref: '#/definitions/ProgrammingLangRequestBody'
            submission:
                type: string
                description: The code submission
                example: print(factorial(5))
            task:
                $ref: '#/definitions/TaskRequestBody'
            username:
                type: string
                description: Username of the user who submitted
                example: coder123
            uuid:
                type: string
                description: UUID of the submission
                example: 123e4567-e89b-12d3-a456-426614174000
        example:
            createdAt: "2024-08-08T10:30:00Z"
            evaluation:
                possibleScore: 100
                receivedScore: 85
                status: completed
                uuid: 123e4567-e89b-12d3-a456-426614174000
            language:
                fullName: Go
                id: go
                monacoId: go
            submission: print(factorial(5))
            task:
                cpu_time_limit_seconds: 0.5
                default_md_statement:
                    input: Ievaddatu pirmajā rindā dotas trīs naturālu skaitļu...
                    notes: Illo officia.
                    output: Izvaddatu vienīgajā rindā jābūt veselam skaitlim...
                    scoring: Asperiores quidem omnis illo.
                    story: Krišjānis ir uzkonstruējis kvadrātveida putekļsūcēju (saīsināti – KP), kas ir neaizstājams palīgs viņa darbnīcas uzkopšanā...
                default_pdf_statement_url: https://dvhk4hiwp1rmf.cloudfront.net/task-pdf-statements/f27386f1b69f3c020335fe7c84316c5da099933832043e6a5820bcdb0cd66a81.pdf
                difficulty_rating: 3
                examples:
                    - input: |
                        5 9 3
                        A....X..B
                        ..X..X.X.
                        .XXX.XX..
                        X.X.X..X.
                        ...XX....
                      md_note: Qui tempora aut ea.
                      output: |
                        10
                    - input: |
                        5 9 3
                        A....X..B
                        ..X..X.X.
                        .XXX.XX..
                        X.X.X..X.
                        ...XX....
                      md_note: Qui tempora aut ea.
                      output: |
                        10
                    - input: |
                        5 9 3
                        A....X..B
                        ..X..X.X.
                        .XXX.XX..
                        X.X.X..X.
                        ...XX....
                      md_note: Qui tempora aut ea.
                      output: |
                        10
                illustration_img_url: https://dvhk4hiwp1rmf.cloudfront.net/task-illustrations/bafcc0aa1b4c56faa44f5f3b2a5abd9529af941eb9a9b10f541b436762d4948a.png
                memory_limit_megabytes: 256
                origin_notes:
                    lv: Uzdevums parādījās Latvijas 37. informātikas olimpiādes (2023./2024. gads) skolas kārtā.
                origin_olympiad: LIO
                published_task_id: kvadrputekl
                task_full_name: Kvadrātveida putekļsūcējs
                visible_input_subtasks:
                    - inputs:
                        - |
                          5 5 3
                          A.X..
                          X.B.X
                          ....X
                          XX...
                          X.XXX
                        - |
                          6 3 1
                          ...
                          XA.
                          .X.
                          X..
                          B..
                          ...
                        - |
                          6 4 2
                          X...
                          .AXX
                          X..X
                          .X.X
                          XXXX
                          X.BX
                      subtask: 1
                    - inputs:
                        - |
                          5 5 3
                          A.X..
                          X.B.X
                          ....X
                          XX...
                          X.XXX
                        - |
                          6 3 1
                          ...
                          XA.
                          .X.
                          X..
                          B..
                          ...
                        - |
                          6 4 2
                          X...
                          .AXX
                          X..X
                          .X.X
                          XXXX
                          X.BX
                      subtask: 1
                    - inputs:
                        - |
                          5 5 3
                          A.X..
                          X.B.X
                          ....X
                          XX...
                          X.XXX
                        - |
                          6 3 1
                          ...
                          XA.
                          .X.
                          X..
                          B..
                          ...
                        - |
                          6 4 2
                          X...
                          .AXX
                          X..X
                          .X.X
                          XXXX
                          X.BX
                      subtask: 1
                    - inputs:
                        - |
                          5 5 3
                          A.X..
                          X.B.X
                          ....X
                          XX...
                          X.XXX
                        - |
                          6 3 1
                          ...
                          XA.
                          .X.
                          X..
                          B..
                          ...
                        - |
                          6 4 2
                          X...
                          .AXX
                          X..X
                          .X.X
                          XXXX
                          X.BX
                      subtask: 1
            username: coder123
            uuid: 123e4567-e89b-12d3-a456-426614174000
        required:
            - uuid
            - submission
            - username
            - createdAt
            - evaluation
            - language
            - task
...
securityDefinitions:
    jwt_header_Authorization:
        type: apiKey
        description: |4-
            **Security Scopes**:
              * `users:read`: Read users
              * `users:write`: Write users
        name: Authorization
        in: header
    jwt_query_Authorization:
        type: apiKey
        description: |4-
            **Security Scopes**:
              * `users:read`: Read users
              * `users:write`: Write users
        name: Authorization
        in: query